lab3.2c edit printing of result

+++ Programming-Lab3.2c/Programming-Lab3.2c.c
@@ -17,7 +17,7 @@ int main()
 		P *= (float)(m - i) / (i + 1);
 	}
 
-	printf("P = %.0f", P);
+	printf("P = %d", (int)P);
 
 	getch();
 }
